fix: annotation location always points to the test file
