+++ packages/runner/src/utils/collect.ts
@@ -1,5 +1,7 @@
+import type { ParsedStack } from '@vitest/utils'
 import type { File, Suite, TaskBase } from '../types/tasks'
 import { processError } from '@vitest/utils/error'
+import { parseSingleStack } from '@vitest/utils/source-map'
 import { relative } from 'pathe'
 import { setFileContext } from '../context'
 
@@ -209,4 +211,15 @@ export function generateFileHash(
   projectName: string | undefined,
 ): string {
   return generateHash(`${file}${projectName || ''}`)
+}
+
+export function findTestFileStackTrace(testFilePath: string, error: string): ParsedStack | undefined {
+  // first line is the error message
+  const lines = error.split('\n').slice(1)
+  for (const line of lines) {
+    const stack = parseSingleStack(line)
+    if (stack && stack.file === testFilePath) {
+      return stack
+    }
+  }
 }
